Fix NPE during audio recording (fix #459)

+++ app/src/main/java/org/kontalk/ui/view/ComposerBar.java
@@ -526,7 +526,7 @@ public class ComposerBar extends RelativeLayout implements
                 Toast.LENGTH_LONG).show();
         }
         finally {
-            if (!canSend)
+            if (!canSend && mRecordFile != null)
                 mRecordFile.delete();
         }
     }
